dENUMinating fan processing
- Transform Fan Processing Types from Enum to Abstract class with subclasses
- Reorganise AirFlowParticle code and move type specific code to types themselves
