+++ src/main/java/com/simibubi/create/content/kinetics/fan/HauntingRecipe.java
@@ -3,20 +3,21 @@ package com.simibubi.create.content.kinetics.fan;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import com.simibubi.create.AllRecipeTypes;
+import com.simibubi.create.content.kinetics.fan.processingTypes.HauntingType;
 import com.simibubi.create.content.processing.recipe.ProcessingRecipe;
 import com.simibubi.create.content.processing.recipe.ProcessingRecipeBuilder.ProcessingRecipeParams;
 
 import net.minecraft.world.level.Level;
 
 @ParametersAreNonnullByDefault
-public class HauntingRecipe extends ProcessingRecipe<FanProcessing.HauntingWrapper> {
+public class HauntingRecipe extends ProcessingRecipe<HauntingType.HauntingWrapper> {
 
 	public HauntingRecipe(ProcessingRecipeParams params) {
 		super(AllRecipeTypes.HAUNTING, params);
 	}
 
 	@Override
-	public boolean matches(FanProcessing.HauntingWrapper inv, Level worldIn) {
+	public boolean matches(HauntingType.HauntingWrapper inv, Level worldIn) {
 		if (inv.isEmpty())
 			return false;
 		return ingredients.get(0)
