+++ src/main/java/com/simibubi/create/content/logistics/filter/ItemAttribute.java
@@ -9,11 +9,13 @@ import java.util.function.BiPredicate;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
+import com.simibubi.create.content.kinetics.fan.processingTypes.HauntingType;
+import com.simibubi.create.content.kinetics.fan.processingTypes.SplashingType;
+
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
 
 import com.simibubi.create.AllRecipeTypes;
-import com.simibubi.create.content.kinetics.fan.FanProcessing;
 import com.simibubi.create.content.logistics.filter.attribute.BookAuthorAttribute;
 import com.simibubi.create.content.logistics.filter.attribute.BookCopyAttribute;
 import com.simibubi.create.content.logistics.filter.attribute.ColorAttribute;
@@ -145,8 +147,8 @@ public interface ItemAttribute {
 		EQUIPABLE(s -> LivingEntity.getEquipmentSlotForItem(s)
 			.getType() != EquipmentSlot.Type.HAND),
 		FURNACE_FUEL(AbstractFurnaceBlockEntity::isFuel),
-		WASHABLE(FanProcessing::isWashable),
-		HAUNTABLE(FanProcessing::isHauntable),
+		WASHABLE(SplashingType::isWashable),
+		HAUNTABLE(HauntingType::isHauntable),
 		CRUSHABLE((s, w) -> testRecipe(s, w, AllRecipeTypes.CRUSHING.getType())
 			|| testRecipe(s, w, AllRecipeTypes.MILLING.getType())),
 		SMELTABLE((s, w) -> testRecipe(s, w, RecipeType.SMELTING)),
